GCP/k6_runner_service/requirements.txt
Flask==3.0.3
google-cloud-storage
gunicorn==22.0.0
